feat: handle config files

+++ src/iguana/algorithms/clas12/ZVertexFilterBindings.cc
@@ -8,7 +8,7 @@ namespace iguana::bindings {
     /// @param [in] algo_idx the algorithm index
     /// @param [in] vz
     /// @param [in,out] out the return value
-    void iguana_clas12_zvertexfilter_filter_(algo_idx_t* algo_idx, int* vz, bool* out)
+    void iguana_clas12_zvertexfilter_filter_(algo_idx_t* algo_idx, float* vz, bool* out)
     {
       *out = *out && dynamic_cast<clas12::ZVertexFilter*>(iguana_get_algo_(algo_idx))->Filter(*vz);
     }
